requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pystache==0.5.4
pytz==2019.3
sqlparse==0.3.1
